requirements.txt
Flask
google-generativeai
markdown-it-py
python-docx
sxtwl
gunicorn